Move common functions in netbox/utils sub-module
  This is done in preparation for splitting up the NetBox connector and client
  to support version specific NetBox clients and to support multiple version of
  NetBox at the same time.

  Also fix some linter nits.
